% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustree.R
\name{add_node_labels}
\alias{add_node_labels}
\title{Add node labels}
\usage{
add_node_labels(
  node_label,
  node_colour,
  node_label_size,
  node_label_colour,
  node_label_nudge,
  allowed
)
}
\arguments{
\item{node_label}{the name of a metadata column for node labels}

\item{node_colour}{either a value indicating a colour to use for all nodes or
the name of a metadata column to colour nodes by}

\item{node_label_size}{size of node label text}

\item{node_label_colour}{colour of node_label text}

\item{node_label_nudge}{numeric value giving nudge in y direction for node
labels}

\item{allowed}{vector of allowed node attributes to use as aesthetics}
}
\description{
Add node labels to a clustering tree plot with the specified aesthetics.
}
\keyword{internal}
