% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustree.R
\name{add_node_points}
\alias{add_node_points}
\title{Add node points}
\usage{
add_node_points(prefix, node_colour, node_size, node_alpha, metadata)
}
\arguments{
\item{prefix}{string indicating columns containing clustering information}

\item{node_colour}{either a value indicating a colour to use for all nodes or
the name of a metadata column to colour nodes by}

\item{node_size}{either a numeric value giving the size of all nodes or the
name of a metadata column to use for node sizes}

\item{node_alpha}{either a numeric value giving the alpha of all nodes or the
name of a metadata column to use for node transparency}

\item{metadata}{data.frame containing metadata on each sample that can be
used as node aesthetics}
}
\description{
Add node points to a clustering tree plot with the specified aesthetics.
}
