% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{aggr_metadata}
\alias{aggr_metadata}
\title{Aggregate metadata}
\usage{
aggr_metadata(node_data, col_name, col_aggr, metadata, is_cluster)
}
\arguments{
\item{node_data}{data.frame containing information about a set of cluster
nodes}

\item{col_name}{the name of the metadata column to aggregate}

\item{col_aggr}{string naming a function used to aggregate the column}

\item{metadata}{data.frame providing metadata on samples}

\item{is_cluster}{logical vector indicating which rows of metadata are in the
node to be summarized}
}
\value{
data.frame with aggregated data
}
\description{
Aggregate a metadata column to get a summarized value for a cluster node
}
\keyword{internal}
