% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{alternate_clustree_layers}
\alias{alternate_clustree_layers}
\title{Alternate clustering tree layers}
\usage{
alternate_clustree_layers(plot, split_by = "res")
}
\arguments{
\item{plot}{A \code{ggplot} object}

\item{split_by}{Name of the column in \code{plot$data} to split layers by}
}
\value{
A modified \code{ggplot} object, see \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} for details
}
\description{
Alternate layers of a \code{ggplot} object representing the nodes and edges of a
clustering tree so that they are plotted in the expected order. Usually used
with a plot created with \code{\link[=overlay_clustree]{overlay_clustree()}}.
}
\details{
The function first searches for layers in the plot created with
\code{\link[=geom_clustree_point]{geom_clustree_point()}} and \code{\link[=geom_clustree_edge]{geom_clustree_edge()}} (exactly one of each).
If these are found the layers are extracted and split according to values in
\code{split_by}. The resulting split layers are then alternated (node, edge, node,
...) and placed back in the original plot object. Any other layers already in
the plot should remain in the same position, except any placed between the
original node and edge layers which will be moved towards the top (later in
the plotting order).

\strong{Limitations}

Because this function modifies the internal structure of a \code{ggplot} object it
has some limitations and assumptions. These mean it may not be compatible
with more complex clustering tree plots, such as when custom data has been
supplied to a layer. It may also cause other unwanted changes to the plot so
should be used with some amount of caution.
}
\examples{
graph <- build_clustree_graph(nba_clusts, pattern = "K(.*)")
graph <- summarise_metadata(graph, PC1 = mean(PC1), PC2 = mean(PC2))
overlay <- overlay_clustree(graph, x_dim = "PC1", y_dim = "PC2")
plot <- plot_clustree(overlay) +
    geom_clustree_point() +
    geom_clustree_edge()
alternate_clustree_layers(plot)
}
