% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{assert_numeric_node_aes}
\alias{assert_numeric_node_aes}
\title{Assert numeric node aesthetics}
\usage{
assert_numeric_node_aes(node_aes_name, prefix, metadata, node_aes,
  node_aes_aggr, min, max)
}
\arguments{
\item{node_aes_name}{name of the node aesthetic to check}

\item{prefix}{string indicating columns containing clustering information}

\item{metadata}{data.frame containing metadata on each sample that can be
used as node aesthetics}

\item{node_aes}{value of the node aesthetic to check}

\item{node_aes_aggr}{aggregation function associated with the node aesthetic}

\item{min}{minimum numeric value allowed}

\item{max}{maximum numeric value allowed}
}
\description{
Raise error if an incorrect set of numeric node parameters has been supplied.
}
