% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph.R
\name{build_clustree_graph}
\alias{build_clustree_graph}
\alias{build_clustree_graph.default}
\alias{build_clustree_graph.matrix}
\alias{build_clustree_graph.data.frame}
\alias{build_clustree_graph.SingleCellExperiment}
\alias{build_clustree_graph.Seurat}
\title{Build clustering tree graph}
\usage{
build_clustree_graph(x, ...)

\method{build_clustree_graph}{default}(x, metadata = NULL, ...)

\method{build_clustree_graph}{matrix}(x, metadata = NULL, ...)

\method{build_clustree_graph}{data.frame}(
  x,
  pattern = NULL,
  prefix = NULL,
  suffix = NULL,
  clust_cols = NULL,
  ...
)

\method{build_clustree_graph}{SingleCellExperiment}(x, exprs = "counts", features = NULL, ...)

\method{build_clustree_graph}{Seurat}(
  x,
  exprs = c("data", "counts", "scale.data"),
  exprs_type = NULL,
  features = NULL,
  ...
)
}
\arguments{
\item{x}{Object containing clustering data}

\item{...}{Arguments used by other methods}

\item{metadata}{data.frame containing metadata about individual samples}

\item{pattern}{A regular expression matching column names for columns
containing clustering data. Should have a single capture group.}

\item{prefix}{Prefix string used to select columns containing clustering
information. Combined with \code{suffix} to create a regular expression.}

\item{suffix}{Suffix string used to select columns containing clustering
information. Combined with \code{prefix} to create a regular expression.}

\item{clust_cols}{Vector of column names used to select columns containing
clustering information.}

\item{exprs}{Source of feature expression information to extract. For
\code{SingleCellExperiment} objects it must be a name in \code{assayNames(x)} and for
\code{Seurat} object it must be the name of an assay slot ("data", "counts" or
"scale.data").}

\item{features}{Vector of feature names to extract.}

\item{exprs_type}{Type of feature expression to extract. For \code{Seurat} objects
it must be a name in \code{Assays(x)}.}
}
\value{
A \code{clustree_graph} object
}
\description{
Construct a clustering tree graph from various objects
}
\details{
\strong{Column selection}

A key part of building a clustering tree graph is selecting columns in an
object which contain clustering information. This can be done in a variety
of ways. The primary method is to specify the \code{pattern} argument. This is a
regular expression containing a single capture group which is applied to
column names. For example \code{pattern = "K(.*)"} would match columns starting
with "K" and capture the rest of the column name. Setting
\code{pattern = "^Res_(.*)_cluster"} would match columns starting with "Res_" and
ending with "_cluster" and extract the part in between.

A simpler alternative to setting \code{pattern} is to use the \code{prefix} and
\code{suffix} arguments. These are pasted together to create a pattern so that
\verb{prefix = "Res", suffix = "Cluster"} would be equivalent to
\verb{pattern = "Res(.*)Cluster}.

Exact column names can also be supplied using the \code{clust_cols} argument.
For example \code{clust_cols = c("K1", "K3", "K5")} would select only those
columns. If \code{clust_cols} is named those names will be used for each
resolution level.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method. Computes nodes and edges and
constructs the graph.

\item \code{matrix}: Method for \code{matrix} objects. Calls default
method.

\item \code{data.frame}: Method for \code{data.frame} objects. Extracts
clusterings and metadata and passes them to the \code{matrix} method.

\item \code{SingleCellExperiment}: Method for \code{SingleCellExperiment} objects.
Extracts \code{colData}, adds features and passes it to the \code{data.frame} method.

\item \code{Seurat}: Method for \code{Seurat} objects.
Extracts \code{metadata}, adds features and passes it to the \code{data.frame} method.
}}

\examples{
build_clustree_graph(nba_clusts, pattern = "K(.*)")
}
\seealso{
\code{\link[=clustree_graph]{clustree_graph()}} for more details on the \code{clustree_graph} object
}
