% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{calc_sc3_stability}
\alias{calc_sc3_stability}
\title{Calculate SC3 stability}
\usage{
calc_sc3_stability(clusterings)
}
\arguments{
\item{clusterings}{numeric matrix containing clustering information, each
column contains clustering at a separate resolution}
}
\value{
matrix with stability score for each cluster
}
\description{
Calculate the SC3 stability index for every cluster at every resolution in a
set of clusterings. The index varies from 0 to 1, where 1 suggests that a
cluster is more stable across resolutions. See \code{\link[=calc_sc3_stability_cluster]{calc_sc3_stability_cluster()}}
for more details.
}
\keyword{internal}
