% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_node_aes_list}
\alias{check_node_aes_list}
\title{Check node aes list}
\usage{
check_node_aes_list(node_aes_list)
}
\arguments{
\item{node_aes_list}{List of node aesthetics}
}
\value{
Corrected node aesthetics list
}
\description{
Warn if node aesthetic names are incorrect
}
\keyword{internal}
