% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustree-package.r
\docType{package}
\name{clustree-package}
\alias{clustree-package}
\title{Clustree}
\description{
Deciding what resolution to use can be a difficult question when
approaching a clustering analysis. One way to approach this problem is to
look at how samples move as the number of clusters increases. This package
allows you to produce clustering trees, a visualisation for interrogating
clusterings as resolution increases.
}
