% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustree_graph.R
\name{clustree_graph}
\alias{clustree_graph}
\alias{as_clustree_graph}
\alias{as_clustree_graph.default}
\alias{as_clustree_graph.tbl_graph}
\alias{is.clustree_graph}
\title{Data structure for manipulating clustering trees}
\usage{
clustree_graph(..., metadata = NULL)

as_clustree_graph(x, ..., metadata = NULL)

\method{as_clustree_graph}{default}(x, ..., metadata = NULL)

\method{as_clustree_graph}{tbl_graph}(x, ..., metadata = NULL)

is.clustree_graph(x)
}
\arguments{
\item{...}{Arguments passed to \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph()}} or conversion
functions}

\item{metadata}{A \code{data.frame} providing information about individual
samples. If a column named \code{.clustree_idx} is not already present it will
be created using the row names. If \code{NULL} an empty \code{tibble} is created.}

\item{x}{Object to convert to a \code{clustree_graph}}
}
\value{
A \code{clustree_graph} object
}
\description{
The \code{clustree_graph} class is an extension to the \code{tbl_graph} class which
adds a table for storing metadata about individual samples. Unlike most
graphs where is node is a sample, nodes in a clustering tree represent
clusters of multiple samples. Storing metadata about the original samples
makes it possible to calculate additional node (or edge) level summaries as
required rather than having to compute these when the graph is created.
}
\details{
A \code{clustree_graph} object should usually be created using the
\code{\link[=build_clustree_graph]{build_clustree_graph()}} function. If using \code{\link[=clustree_graph]{clustree_graph()}} directly then
both the \code{nodes} and \code{edges} tables must include a column named
"\code{.clustree_indices}" which is a list of vectors giving the indices in
the \code{.clustree_idx} column of \code{metadata} matching the samples associated
with each node/edge.

The metadata table is stored as a graph attribute accessible with
\code{\link[igraph:graph_attr]{igraph::graph_attr()}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{as_clustree_graph}: Default method. Tries to call
\code{\link[tidygraph:tbl_graph]{tidygraph::as_tbl_graph()}} on the input before converting.

\item \code{as_clustree_graph}: Method for \code{tbl_graph} objects. Stores \code{metadata}
and appends the class.
}}

\examples{
nodes <- tibble::tibble(
    node = c("A", "B"),
    .clustree_indices = list(c("1", "2"), c("1"))
)

edges <- tibble::tibble(
    from = "A",
    to   = "B",
    .clustree_indices = list(c("1"))
)

metadata <- data.frame(
    .clustree_idx = c("1", "2"),
    value = c("Example1", "Example2")
)

clustree_graph(nodes = nodes, edges = edges, metadata = metadata)
}
\seealso{
\code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph()}} for the \code{tbl_graph} object and
\code{\link[=build_clustree_graph]{build_clustree_graph()}} for building \code{clustree_graph} objects from various
inputs
}
