% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_clustree_edges}
\alias{geom_clustree_edges}
\title{Clustree edges}
\usage{
geom_clustree_edges(mapping = NULL, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}}

\item{...}{Other argument passed on to \code{\link[ggraph:geom_edge_link]{ggraph::geom_edge_link()}}}
}
\description{
The clustree edges geom is used to add lines representing edges in a
graph to a clustering tree plot. It is a minimal wrapper around
\code{\link[ggraph:geom_edge_link]{ggraph::geom_edge_link()}} that sets some default aesthetics.
}
\section{Default aesthetics}{

\itemize{
\item \code{edge_colour = count}
\item \code{edge_alpha = in_prop}
\item \code{width = 1.5}
\item \code{arrow = grid::arrow(length = grid::unit(7.5, "points"), ends = "last")}
\item \code{end_cap = ggraph::circle(15.5, "points")}
}
}

\examples{
graph <- build_clustree_graph(nba_clusts, pattern = "K(.*)")
plot_clustree(graph) +
    geom_clustree_edges()
}
\seealso{
\code{\link[=plot_clustree]{plot_clustree()}} for creating the plot object and
\code{\link[=geom_clustree_nodes]{geom_clustree_nodes()}} and \code{\link[=geom_clustree_text]{geom_clustree_text()}} for other clustering tree
geoms
}
