% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_clustree_nodes}
\alias{geom_clustree_nodes}
\title{Clustree nodes}
\usage{
geom_clustree_nodes(mapping = NULL, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}}

\item{...}{Other argument passed on to \code{\link[ggraph:geom_node_point]{ggraph::geom_node_point()}}}
}
\description{
The clustree nodes geom is used to add points representing each node in the
graph to a clustering tree plot. It is a minimal wrapper around
\code{\link[ggraph:geom_node_point]{ggraph::geom_node_point()}} that sets some default aesthetics.
}
\section{Default aesthetics}{

\itemize{
\item \code{colour = res}
\item \code{size = size}
}
}

\examples{
graph <- build_clustree_graph(nba_clusts, pattern = "K(.*)")
plot_clustree(graph) +
    geom_clustree_nodes()
}
\seealso{
\code{\link[=plot_clustree]{plot_clustree()}} for creating the plot object and
\code{\link[=geom_clustree_text]{geom_clustree_text()}} and \code{\link[=geom_clustree_edges]{geom_clustree_edges()}} for other clustering tree
geoms. See \code{\link[ggraph:geom_node_point]{ggraph::geom_node_point()}} for details on the underlying geom.
}
