% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_clustree_point}
\alias{geom_clustree_point}
\title{Clustree node points}
\usage{
geom_clustree_point(
  mapping = NULL,
  data = NULL,
  position = "identity",
  show.legend = NA,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}
or \code{\link[ggplot2:aes_]{ggplot2::aes_()}}. By default x and y are mapped to x and y in
the node data.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
The clustree points geom is used to add points representing each node in the
graph to a clustering tree plot. It is equivalent to
\code{\link[ggraph:geom_node_point]{ggraph::geom_node_point()}} or \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} but sets some default
aesthetics.
}
\section{Aesthetics}{

\code{geom_clustree_point} understand the following aesthetics. Bold aesthetics
are automatically set, but can be overridden.
\itemize{
\item \strong{x}
\item \strong{y}
\item alpha
\item \strong{colour = res}
\item fill
\item shape
\item \strong{size = size}
\item stroke
\item filter
}
}

\examples{
graph <- build_clustree_graph(nba_clusts, pattern = "K(.*)")
plot_clustree(graph) +
    geom_clustree_point()
}
\seealso{
\code{\link[=plot_clustree]{plot_clustree()}} for creating the plot object and
\code{\link[=geom_clustree_text]{geom_clustree_text()}} and \code{\link[=geom_clustree_edge]{geom_clustree_edge()}} for other clustering tree
geoms. See \code{\link[ggraph:geom_node_point]{ggraph::geom_node_point()}} for details on the underlying geom.
}
