% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_clustree_text}
\alias{geom_clustree_text}
\title{Clustree text}
\usage{
geom_clustree_text(mapping = NULL, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}}

\item{...}{Other argument passed on to \code{\link[ggraph:geom_node_text]{ggraph::geom_node_text()}}}
}
\description{
The clustree text geom is used to add text labelling each node in a
graph to a clustering tree plot. It is a minimal wrapper around
\code{\link[ggraph:geom_node_text]{ggraph::geom_node_text()}} that sets some default aesthetics.
}
\section{Default aesthetics}{

\itemize{
\item \code{label = cluster}
\item \code{size = 3}
}
}

\examples{
graph <- build_clustree_graph(nba_clusts, pattern = "K(.*)")
plot_clustree(graph) +
    geom_clustree_text()
}
\seealso{
\code{\link[=plot_clustree]{plot_clustree()}} for creating the plot object and
\code{\link[=geom_clustree_nodes]{geom_clustree_nodes()}} and \code{\link[=geom_clustree_edges]{geom_clustree_edges()}} for other clustering tree
geoms. See \code{\link[ggraph:geom_node_text]{ggraph::geom_node_text()}} for details on the underlying geom.
}
