% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_metadata_point}
\alias{geom_metadata_point}
\title{Clustree metadata points}
\usage{
geom_metadata_point(
  mapping = NULL,
  data = NULL,
  position = "identity",
  show.legend = NA,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
The metadata points geom is used to add points representing individual
samples to a clustering tree plot using the "overlay" layout. It is
equivalent to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.
}
\section{Aesthetics}{

\code{geom_metadata_point} understands the following aesthetics. Bold aesthetics
are automatically set, but can be overridden.
\itemize{
\item \strong{x}
\item \strong{y}
\item alpha
\item colour
\item fill
\item shape
\item size
\item stroke
\item filter
}
}

\examples{
graph <- build_clustree_graph(nba_clusts, pattern = "K(.*)")
graph <- summarise_metadata(graph, PC1 = mean(PC1), PC2 = mean(PC2))
plot_clustree(graph, layout = "overlay", x_dim = "PC1", y_dim = "PC2") +
    geom_metadata_point()
}
\seealso{
\code{\link[=plot_clustree]{plot_clustree()}} for creating the plot object and
\code{\link[=geom_clustree_point]{geom_clustree_point()}}, \code{\link[=geom_clustree_text]{geom_clustree_text()}} and \code{\link[=geom_clustree_edge]{geom_clustree_edge()}}
for other clustering tree geoms. See \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} for details on
the underlying geom.
}
