% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{get_tree_edges}
\alias{get_tree_edges}
\title{Get tree edges}
\usage{
get_tree_edges(clusterings, prefix)
}
\arguments{
\item{clusterings}{numeric matrix containing clustering information, each
column contains clustering at a separate resolution}

\item{prefix}{string indicating columns containing clustering information}
}
\value{
data.frame containing edge information
}
\description{
Extract the edges from a set of clusterings
}
\keyword{internal}
