% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{get_tree_nodes}
\alias{get_tree_nodes}
\title{Get tree nodes}
\usage{
get_tree_nodes(clusterings, prefix, metadata, node_aes_list)
}
\arguments{
\item{clusterings}{numeric matrix containing clustering information, each
column contains clustering at a separate resolution}

\item{prefix}{string indicating columns containing clustering information}

\item{metadata}{data.frame containing metadata on each sample that can be
used as node aesthetics}

\item{node_aes_list}{nested list containing node aesthetics}
}
\value{
data.frame containing node information
}
\description{
Extract the nodes from a set of clusterings and add relevant attributes
}
