% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_clustree}
\alias{layout_clustree}
\alias{layout_clustree.default}
\alias{layout_clustree.clustree_graph}
\title{Layout clustering tree}
\usage{
layout_clustree(x, ...)

\method{layout_clustree}{default}(x, ...)

\method{layout_clustree}{clustree_graph}(x, layout = c("tree", "sugiyama"), ...)
}
\arguments{
\item{...}{Arguments used by other methods}

\item{layout}{String specifying the "tree" or "sugiyama" layout. See
\code{\link[igraph:layout_as_tree]{igraph::layout_as_tree()}} and \code{\link[igraph:layout_with_sugiyama]{igraph::layout_with_sugiyama()}} for details.}
}
\value{
A \code{layout_ggraph} object, see \code{\link[ggraph:ggraph]{ggraph::layout_ggraph()}} for details
}
\description{
Create a layout for a clustering tree
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method. Tries to call
\code{\link[=as_clustree_graph]{as_clustree_graph()}} on the input before creating the layout.

\item \code{clustree_graph}: Method for \code{clustree_graph} objects. Creates
the layout.
}}

\examples{
graph <- build_clustree_graph(nba_clusts, pattern = "K(.*)")
layout <- layout_clustree(graph)
}
\seealso{
\code{\link[ggraph:ggraph]{ggraph::layout_ggraph()}} for details on the \code{layout_ggraph} object
and layouts in general
}
