% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_overlay}
\alias{layout_overlay}
\title{Layout overlay}
\usage{
layout_overlay(graph, x_dim, y_dim)
}
\arguments{
\item{graph}{Object containing a clustering tree graph}

\item{x_dim}{Name of the node column to be used as the x dimension}

\item{y_dim}{Name of the node column to be used as the y dimension}
}
\value{
A \code{layout_ggraph} object, see \code{\link[ggraph:ggraph]{ggraph::layout_ggraph()}} for details
}
\description{
Create a layout for a clustering tree by overlaying nodes onto given
dimensions. Usually used by setting \code{layout = "overlay"} in \code{\link[=plot_clustree]{plot_clustree()}}
rather than by calling this function directly.
}
\examples{
graph <- build_clustree_graph(nba_clusts, pattern = "K(.*)")
graph <- summarise_metadata(graph, PC1 = mean(PC1), PC2 = mean(PC2))
layout_overlay(graph, x_dim = "PC1", y_dim = "PC2")
}
\seealso{
\code{\link[ggraph:ggraph]{ggraph::layout_ggraph()}} for details on the \code{layout_ggraph} object
and layouts in general
}
