% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{overlay_clustree}
\alias{overlay_clustree}
\alias{overlay_clustree.default}
\alias{overlay_clustree.clustree_graph}
\title{Overlay clustering tree}
\usage{
overlay_clustree(x, ...)

\method{overlay_clustree}{default}(x, ...)

\method{overlay_clustree}{clustree_graph}(x, x_dim, y_dim, ...)
}
\arguments{
\item{x}{Object containing a clustering tree graph}

\item{...}{Arguments used by other methods}

\item{x_dim}{Name of the node column to be used as the x dimension}

\item{y_dim}{Name of the node column to be used as the y dimension}
}
\value{
A \code{layout_ggraph} object, see \code{\link[ggraph:ggraph]{ggraph::layout_ggraph()}} for details
}
\description{
Create a layout for a clustering tree by overlaying nodes onto given
dimensions
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method. Tries to call
\code{\link[=as_clustree_graph]{as_clustree_graph()}} on the input before creating the layout.

\item \code{clustree_graph}: Method for \code{clustree_graph} objects. Creates
the layout.
}}

\examples{
graph <- build_clustree_graph(nba_clusts, pattern = "K(.*)")
graph <- summarise_metadata(graph, PC1 = mean(PC1), PC2 = mean(PC2))
overlay <- overlay_clustree(graph, x_dim = "PC1", y_dim = "PC2")
}
\seealso{
\code{\link[ggraph:ggraph]{ggraph::layout_ggraph()}} for details on the \code{layout_ggraph} object
and layouts in general
}
