% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustree_overlay.R
\name{overlay_node_points}
\alias{overlay_node_points}
\title{Overlay node points}
\usage{
overlay_node_points(
  nodes,
  x_value,
  y_value,
  node_colour,
  node_size,
  node_alpha
)
}
\arguments{
\item{nodes}{data.frame describing nodes}

\item{x_value}{column of nodes to use for the x position}

\item{y_value}{column of nodes to use for the y position}

\item{node_colour}{either a value indicating a colour to use for all nodes or
the name of a metadata column to colour nodes by}

\item{node_size}{either a numeric value giving the size of all nodes or the
name of a metadata column to use for node sizes}

\item{node_alpha}{either a numeric value giving the alpha of all nodes or the
name of a metadata column to use for node transparency}
}
\description{
Overlay clustering tree nodes on a scatter plot with the specified
aesthetics.
}
