% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustree.r
\name{theme_clustree}
\alias{theme_clustree}
\title{clustree theme}
\usage{
theme_clustree(base_size = 14, base_family = "")
}
\arguments{
\item{base_size}{overall font size}

\item{base_family}{base font family}
}
\description{
Default theme used for plotting clustering trees
}
\examples{
qplot(1:10, (1:10) ^ 2) + theme_clustree()

}
