% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph.R
\name{build_clustree_graph}
\alias{build_clustree_graph}
\alias{build_clustree_graph.default}
\alias{build_clustree_graph.matrix}
\alias{build_clustree_graph.data.frame}
\title{Build clustering tree graph}
\usage{
build_clustree_graph(x, ...)

\method{build_clustree_graph}{default}(x, metadata = NULL)

\method{build_clustree_graph}{matrix}(x, metadata = NULL, ...)

\method{build_clustree_graph}{data.frame}(x, pattern = "(.*)", ...)
}
\arguments{
\item{x}{Object containing clustering data}

\item{...}{other}

\item{metadata}{data.frame metadata about individual samples}

\item{pattern}{A regular expression matching column names for columns
containing clustering data. Should have a single capture group.}
}
\value{
A \code{clustree_graph} object
}
\description{
Construct a clustering tree graph from various objects
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method. Computes nodes and edges and
constructs the graph.

\item \code{matrix}: Method for \code{matrix} objects. Calls default
method.

\item \code{data.frame}: Method for \code{data.frame} objects. Extracts
clusterings and metadata and passes them to other methods.
}}

\examples{
build_clustree_graph(nba_clusts, pattern = "K(.*)")
}
\seealso{
\code{\link[=clustree_graph]{clustree_graph()}} for more details on the \code{clustree_graph} object
}
