% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustree.R
\name{clustree}
\alias{clustree}
\alias{clustree_overlay}
\title{Clustering tree all-in-one functions}
\usage{
clustree(x, ...)

clustree_overlay(x, x_value, y_value, ...)
}
\arguments{
\item{x}{Object containing clustering data}

\item{...}{Arguments passed to \code{\link[=build_clustree_graph]{build_clustree_graph()}} and
\code{\link[=plot_clustree]{plot_clustree()}}. Any additional arguments not used by these functions will
be ignored with a warning.}

\item{x_value}{Name of a numeric metadata column to use as the x-axis}

\item{y_value}{Name of a numeric metadata column to use as the y-axis}
}
\value{
A \code{ggraph} object, see \code{\link[ggraph:ggraph]{ggraph::ggraph()}} for details
}
\description{
All-in-one clustering tree plotting functions. In previous versions of
\strong{{clustre}} these were the default (and only) interface. They are now
soft-deprecated with limited functionality and it is recommended to instead
use \code{\link[=build_clustree_graph]{build_clustree_graph()}} to construct the clustering tree graph, followed
by \code{\link[=plot_clustree]{plot_clustree()}} to create the plot.
}
\details{
This functions are maintained for compatibility with code developed for
older versions of \strong{{clustree}}, however the functionality is more limited
and you should receive a warning if you attempt to use arguments that are no
longer supported. It is recommended that code is updated to use
\code{\link[=build_clustree_graph]{build_clustree_graph()}} and \code{\link[=plot_clustree]{plot_clustree()}} directly.

The supplied \code{x} object is based to \code{\link[=build_clustree_graph]{build_clustree_graph()}} to create the
clustering tree graph which is plotted using \code{\link[=plot_clustree]{plot_clustree()}} with the
default \code{\link[=geom_clustree_edge]{geom_clustree_edge()}}, \code{\link[=geom_clustree_point]{geom_clustree_point()}} and
\code{\link[=geom_clustree_text]{geom_clustree_text()}} geoms. For \code{clustree_overlay()} an additional step
involving \code{\link[=summarise_metadata]{summarise_metadata()}} is used to add additional metadata columns
from \code{x_value} and \code{y_value}.
}
\examples{
clustree(nba_clusts, prefix = "K")

clustree_overlay(nba_clusts, prefix = "K", x_value = "PC1", y_value = "PC2")
}
