% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustree.R
\name{clustree}
\alias{clustree}
\alias{clustree.matrix}
\alias{clustree.data.frame}
\alias{clustree.SingleCellExperiment}
\alias{clustree.seurat}
\title{Plot a clustering tree}
\usage{
clustree(x, ...)

\method{clustree}{matrix}(x, prefix, suffix = NULL, count_filter = 0,
  prop_filter = 0.1, metadata = NULL, node_colour = prefix,
  node_colour_aggr = NULL, node_size = "size", node_size_aggr = NULL,
  node_size_range = c(4, 15), node_alpha = 1, node_alpha_aggr = NULL,
  node_text_size = 3, scale_node_text = FALSE, node_text_colour = "black",
  edge_width = 1.5, edge_arrow = TRUE, layout = c("tree", "sugiyama"),
  ...)

\method{clustree}{data.frame}(x, prefix, ...)

\method{clustree}{SingleCellExperiment}(x, prefix, exprs = "counts", ...)

\method{clustree}{seurat}(x, prefix = "res.", exprs = c("data", "raw.data",
  "scale.data"), ...)
}
\arguments{
\item{x}{object containing clustering data}

\item{...}{extra parameters passed to other methods}

\item{prefix}{string indicating columns containing clustering information}

\item{suffix}{string at the end of column names containing clustering
information}

\item{count_filter}{count threshold for filtering edges in the clustering
graph}

\item{prop_filter}{proportion threshold for filtering edges in the clustering
graph}

\item{metadata}{data.frame containing metadata on each sample that can be
used as node aesthetics}

\item{node_colour}{either a value indicating a colour to use for all nodes or
the name of a metadata column to colour nodes by}

\item{node_colour_aggr}{if \code{node_colour} is a column name than a string
giving the name of a function to aggregate that column for samples in each
cluster}

\item{node_size}{either a numeric value giving the size of all nodes or the
name of a metadata column to use for node sizes}

\item{node_size_aggr}{if \code{node_size} is a column name than a string
giving the name of a function to aggregate that column for samples in each
cluster}

\item{node_size_range}{numeric vector of length two giving the maximum and
minimum point size for plotting nodes}

\item{node_alpha}{either a numeric value giving the alpha of all nodes or the
name of a metadata column to use for node transparency}

\item{node_alpha_aggr}{if \code{node_aggr} is a column name than a string
giving the name of a function to aggregate that column for samples in each
cluster}

\item{node_text_size}{numeric value giving the size of node labels if
\code{scale_node_text} is \code{FALSE}}

\item{scale_node_text}{logical indicating whether to scale node labels along
with the node size}

\item{node_text_colour}{colour value for node labels}

\item{edge_width}{numeric value giving the width of plotted edges}

\item{edge_arrow}{logical indicating whether to add an arrow to edges}

\item{layout}{character specifying the "tree" or "sugiyama" layout, see
\code{\link[igraph:layout_as_tree]{igraph::layout_as_tree()}} and \code{\link[igraph:layout_with_sugiyama]{igraph::layout_with_sugiyama()}} for details}

\item{exprs}{source of gene expression information to use as node aesthetics,
for \code{SingleCellExperiment} objects it must be a name in
\code{\link[SummarizedExperiment:assayNames]{SummarizedExperiment::assayNames()}}, for a \code{seurat} object it must be one of
\code{data}, \code{raw.data} or \code{scale.data}}
}
\value{
\link[ggplot2:ggplot]{ggplot2::ggplot} object containing a clustering tree
}
\description{
Creates a plot of a clustering tree showing the relationship between
clusterings at different resolutions.
}
\details{
\strong{Data sources}

Plotting a clustering tree requires information about which cluster each
sample has been assigned to at different resolutions. This information can
be supplied in various forms, as a matrix, data.frame or more specialised
object. In all cases the object provided must contain numeric columns with
the naming structure \code{PXS} where \code{P} is a prefix indicating that the column
contains clustering information, \code{X} is a numeric value indicating the
clustering resolution and \code{S} is any additional suffix to be removed. For
\code{SingleCellExperiment} objects this information must be in the \code{colData} slot
and for \code{Seurat} objects it must be in the \code{meta.data} slot. For all objects
except matrices any additional columns can be used as aesthetics, for
matrices an additional metadata data.frame can be supplied if required.

\strong{Filtering}

Edges in the graph can be filtered by adjusting the \code{count_filter} and
\code{prop_filter} parameters. The \code{count_filter} removes any edges that represent
less than that number of samples, while the \code{prop_filter} removes edges that
represent less than that proportion of cells in the node it points towards.

\strong{Node aesthetics}

The aesthetics of the plotted nodes can be controlled in various ways. By
default the colour indicates the clustering resolution, the size indicates
the number of samples in that cluster and the transparency is set to 100%.
Each of these can be set to a specific value or linked to a supplied metadata
column. For a \code{SingleCellExperiment} or \code{Seurat} object the names of genes
can also be used. If a metadata column is used than an aggregation function
must also be supplied to combine the samples in each cluster. This function
must take a vector of values and return a single value.
}
\examples{
data(iris_clusts)
clustree(iris_clusts, prefix = "K")

}
