% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_clustree}
\alias{plot_clustree}
\alias{plot_clustree.default}
\alias{plot_clustree.clustree_graph}
\alias{plot_clustree.layout_ggraph}
\title{Plot clustering tree}
\usage{
plot_clustree(x, ...)

\method{plot_clustree}{default}(x, ...)

\method{plot_clustree}{clustree_graph}(x, ...)

\method{plot_clustree}{layout_ggraph}(x, ...)
}
\arguments{
\item{...}{Arguments used by other methods}
}
\value{
A \code{ggplot} object, see \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} for details
}
\description{
Plot a clustering tree
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method. Tries to call
\code{\link[=as_clustree_graph]{as_clustree_graph()}} on the input before creating the layout.

\item \code{clustree_graph}: Method for \code{clustree_graph} objects. Calls
\code{\link[=layout_clustree]{layout_clustree()}} on the input before creating the plot.

\item \code{layout_ggraph}: Method for \code{layout_ggraph} objects. Creates
the plot.
}}

\examples{
graph <- build_clustree_graph(nba_clusts, pattern = "K(.*)")
plot_clustree(graph)
}
